// graphql/generated/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
};
// ...
export type Subtask = {
  __typename?: 'Subtask';
  completed: Scalars['Boolean']['output'];
  deadline: Scalars['DateTime']['output'];
  description: Scalars['String']['output'];
  id: Scalars['Int']['output'];
  name: Scalars['String']['output'];
  responsible: User;
};

export type Task = {
  __typename?: 'Task';
  id: Scalars['Int']['output'];
  subtasks: Array<Subtask>;
  title: Scalars['String']['output'];
};

export type User = {
  __typename?: 'User';
  email: Scalars['String']['output'];
  id: Scalars['Int']['output'];
  tasks?: Maybe<Array<Task>>;
  username: Scalars['String']['output'];
};